Fix: 코드리뷰 반영 — LS 호출 견고화·금액 BigDecimal
PR #18 리뷰 반영.

- LS RestClient 공용 빈(LsClientConfig)으로 통합 + connect/read 타임아웃 추가
- LS 업스트림 실패(401 재시도 소진·타임아웃 등)를 EXTERNAL_API_ERROR(502)로 변환
  (raw 예외가 catch-all 500으로 새는 것 방지)
- 현재가 금액 필드 long → BigDecimal(가이드라인·해외 소수점 시세 대비)
- 등락율 파싱 실패를 묵살하지 않고 경고 로그
- DevController: ClassPathResource.getContentAsString으로 스트림 누수 제거
- LsTokenProviderIT: LS 토큰 형식 가정(startsWith eyJ) 제거

+++ ledger-api/src/main/java/com/pocketstock/ledger/trading/dto/StockPriceResponse.java
@@ -4,15 +4,16 @@ import java.math.BigDecimal;
 
 /**
  * 국내 현재가 응답. changePrice·changeRate는 sign(전일대비구분) 적용 후 부호 포함.
+ * ※ 금액 필드는 BigDecimal(가이드라인 + 해외 소수점 시세 대비). volume은 카운트라 long.
  */
 public record StockPriceResponse(
         String stockCode,
-        long currentPrice,
-        long changePrice,
+        BigDecimal currentPrice,
+        BigDecimal changePrice,
         BigDecimal changeRate,
-        long highPrice,
-        long lowPrice,
-        long openPrice,
+        BigDecimal highPrice,
+        BigDecimal lowPrice,
+        BigDecimal openPrice,
         long volume
 ) {
 }
